'Add' links for sectors should have rel="nofollow"
When in production, we found that search crawlers would attempt to
follow each and every link available to them. In the case of the
licence-finder, this meant that they would traverse the entire sector
hierarchy and attempt to add all possible sectors to their search.

Rather than letting this happen, it's easier to notify them to not
bother following (nor indexing) this content.