Add check for new content before checking removal
There was a timing issue where sometimes the page wouldn't have updated
fast enough for the new content to be displayed and the old content
removed.

Capybara will wait for new content to appear on the page (instead of
having an explicit 'wait for' command). So if we add an assertion that
the new content is there capybara will wait for the page to have updated
before running the assertion that the old content has been removed.