Upgrade Slimmer, explicitly set 'wrapper' layout
This should be a no-op change to app behaviour. The Slimmer default was 'wrapper'
before the upgrade, and is not explicitly set.

Slimmer 9.0.0 switched to `core_layout` [1] as the default layout, which is the
modern/recommend layout. Apps not yet using `core_layout` should be explicitly
marked as such, so we can easily identify them as needing updating.

I had a quick look at what would be required to use a modern layout;

- `header_footer_only`:
 - Needs a `related` component
 - button styling
 - some formatting for "Popular licences" list
- `core_layout`:
 - the above, plus...
 - breadcrumbs
 - heading (and subheading) styling

[1] https://github.com/alphagov/slimmer/pull/136