Use ruby 2.2.3
Need to upgrade kgio and raindrops (dependencies of unicorn) to latest versions
that support ruby 2.2.  Need to add test-unit as a specific dependency as
it's no longer part of the std-lib.  When we go up to a newer rails we can
remove this explicit dependency (rails will add mini-test for us).  Also
removed the ruby_debug19 reference from gemfile as it's not useful on ruby 2.x.