Update uncontroversial GDS gems
This means plek, slimmer, govuk-content-schema-test-helpers

We'll do gds-api-adapters in a single commit later as we'll switch it
over into hash mode and raise on missing mode which could be a big
change.  Similarly the jump on govuk_frontend_toolkit is large and
there may be frontend changes we have to check.