Upgrade gds-api-adapters and set to always raise mode
We can't upgrade to the newest because tire relies on rest-client < 2
but we can still go to the version that introduces the
always_raise_for_not_found deprecation warning, which we enable and
rework the code to explicitly deal with HTTPGone and HTTPNotFound
errors.

To make it easier to understand which api adapters we use we move the
content_api and panopticon_registerer references used throughout the
project into the Services module.