Upgrade to ruby 2.3.1
This introduces a warning to the test run from rest-client about using the
Net::HTTP interface.  This comes from deep within JSON parsing so there's
nothing we can really do about it.  The API for RestClient and Net::HTTP
must have changed between ruby versions.  Upgrading GdsApiAdapters and/or
RestClient might remove this warning (but we can't do that until we remove
tire).