Remove Capybara.automatic_reload from integration specs
This was at the top of a couple of integration specs which means it would
act as a global capybara option for any specs run after the spec that
declared it has been required.  If we need this we should enable it
only for the specs that need it, using metadata and a config.before
defined in a support file.  However, removing the line doesn't seem to
make any specs fail, so we'll just drop it entirely.