Use match_array instead of eq where order isn't important
We noticed intermittent spec failures on CI where the order of the elements
on the page wasn't as specified in the spec so it would fail.  The order
doesn't appear to actually be important in terms of what the spec is
testing, so we use match_array instead which is order-agnostic.