Replace Content API call with Rummager
We used to get licences from the Content API by their GDS IDs. This is
basically a search call that the Content Store does not support.

Since we are retiring the Content API and replace it with modern
services and APIs, we decided that using Rummager to fetch licence
metadata (licence identifiers and short descriptions) is an acceptable
replacement.

This commit replaces calls to `Services.content_api.licences_for_ids`
with a call to search in Rummager.