Use multi_match when querying Elasticsearch
Users are currently unable to search for activity or business types using terms that contain slashes, as this results in a server error.

This is because we are using the [query string query type][1] to perform the search. This allows the user to use the full Lucene syntax to perform complex queries by combining terms and operators. However, it means that if the user wants to use an operator as part of a term they need to escape it. Elasticsearch is currently trying to interpret the slash as an operator and is not able to do so in a legal way, so returns an error.

By switching to use the [multi_match query type][2], Elasticsearch will treat the query as a simple term match against the specified fields. This means that the user will no longer be able to use Lucene query syntax, and that they will no longer need to escape slashes to get a valid response.

As the search field in the licence finder is only expected to be used for simple keyword matching, I believe this is a valid trade off to make.

[1]: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html
[2]: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-multi-match-query.html