Remove Lucene escaping logic
Now that we are using a multi_match query, the string the user provides will not be parsed as a Lucene query so there is no need to escape special characters.

The downcase_ending_keywords method wasn’t actually being called from anywhere outside of tests, so that can definitely be removed.