Change 'string' and 'multi_field' types in elasticsearch index
'string' is now 'keyword' (for exact matching) or 'text' (for partial
matching).  There's a simple rule of thumb to update things: if it's
analysed it should be 'text', otherwise it should be 'keyword'.

Elasticsearch 5 can use the 'string' type, but not when there is a
'boost', which the 'activities' field has.  Rather than just change
that one field to 'text', it seemed better to totally remove
the (deprecated) 'string' type.

The other change is that 'multi_field' has been removed, with the
'fields' hash (which is already there) doing the same thing.  I needed
to pick a type for the 'title' field, so I went for 'text' as that's
the same type as all of its subfields.