Make schema definition ES6 compatible
When moving to ES5, we changed the fields from "string" fields to
"text" and "keyword" fields.  This made all of the "analyzed" /
"not_analyzed" settings redundant, but we didn't realise it at the
time: https://www.elastic.co/blog/strings-are-dead-long-live-strings

ES6 removes the "analyzed" / "not_analyzed" options, so we should make
the suggested change.