Eager and auto load files in lib directory
This is done off the back of a short incident that occurred today where
the application started erroring due to a constant (Search) being
unavailable.

This was caused by https://github.com/alphagov/licence-finder/commit/db8c489b2bae147774f79826fc6ea1d3b4b21818
where search used to be set-up by a strange global variable via an
initializer file. Once this was changed search was no longer available
unless the file was required (as was the case in the test environment)
and would error.

To simplify the differences between test and production I've set up all
of the lib files to be eager and auto loaded so that the
dev/test/production files are the same. As part of this I've removed the
various require statements as these are no longer needed.