Pin to Elasticsearch 6
The gem major versions correlate with the version of Elasticsearch.
Since we are using Elasticsearch 6 we should use version 6 of this gem.

Despite the version number being smaller 6.8.2 is actually a much newer
version of this gem than 7.0.0 and resolves a number of issues with
Faraday compatibility. Previously Elasticsearch was failing to load the
correct Faraday constant when there was a problem with Elasticsearch
communication.