Use core_layout template
Licence finder is the last standing application that uses a set of assets from govuk_frontend_toolkit. Since we're not actively working on this app but we don't want this dependency to stop us from dropping govuk_frontend_toolkit in static we port the required styles from static to here and map a few variables so we don't need to call them from govuk_frontend_toolkit.