Install rubocop-govuk 4.0.0.pre.1
This is to test the pre-release out of 4.0.0. Further details are in:
https://github.com/alphagov/rubocop-govuk/issues/129

I've also moved the rubocop-govuk gem into development/test since it
shouldn't be needed as part of production. I've also set the gem as
require false as per the README recommendation since the code of it is
not needed outside of running rubocop.