Remove CONTRIBUTING.md
This repo does not have many contributions, every link in this doc is
out of date and there is nothing repo specific about this file. It seems
that it is simpler to just remove this file than to update it, as
updating it would likely mean information just starting to fall out of
date again.