Add a "browse licences" page which lists all licences
This is to answer a question from BEIS about which licences actually
exist in the licence-finder tool.

I considered adding a rake task to export this information. However, it
takes less than 1 second to compute everything (on my local machine).
And it's not sensitive. So it may as well be a public facing page.

This will also mean we don't have to get a developer to re-run a rake
task if someone wants this information in the future. We can just tell
them to go to www.gov.uk/licence-finder/browse-licences and they'll see
them all.

I've got it to serve the X-Robots-Tag header to stop search engines from
indexing this bit of the site - it can be our little secret.

Licence-finder pages are all cached for 30 mintues (this default is set
in ApplicationController), so there should be no risk of this being
accidentally expensive.