Add a route for licence-finder/browse-licences
For historical reasons, /licence-finder is an "exact" route with a
backend_id of "frontend".

All the other /licence-finder routes are prefix routes, but since they
don't share a common prefix, they all have to be presented to router-api
separately. We present them to router-api by presenting them to
publishing-api, which presents them to content-store, which presents
them to router-api. What do you mean it sounds complicated? Pfff

Since I want to add a new route (licence-finder/browse-licences), I need
to present this to the system using the rake task.

I've added this in the same way as the others. The UUID is meaningless -
I generated it by running `uuidgen | tr '[:upper:]' '[:lower:]'`.