Rake task to add Costs Lawyers Practising Certificate
Colleagues from BEIS have asked us to add this licence to
licence-finder.

It will point to a start page on GOV.UK that links off to the CLSB's
website (i.e. it's not a new licence on licensify).

Apparently the CLSB reached out in April to ask for this, but it's only
just made it into our awareness. It's probably still not that much of a
priority, but it's a good opportunity to work out "how do we actually
add licences to licence-finder?".

From what I've been able to tell from the git history, we've never added
a new licence to licence-finder since the initial data import in 2012 🙈

The closest thing I could find was commit 7c46f89d (from 2013) which
fixes some bits of data which were incorrect in the initial import.

Because we haven't done this in the last 8 years, I had to do quite a
bit of archaeology to work out how everything works. People did not
write good commit messages back in those days. Thankfully we live in a
more civilised age.

The main interesting thing is the format of `gds_id` - all the numbers
in the databasse are of the form `\d{4}-[1-8]-\d'. Looks suspiciously
like it's formatted to be meaningful, right? But what does it mean?

It turns out the first part is a reference to the "Local government
services list (LGSL)":

https://standards.esd.org.uk/?uri=list%2FenglishAndWelshServices

These are things like "Abnormal load notification (785)". It looks like
an old attempt at a register of services, but it hasn't been maintained.
Even when the initial licence import happened, most of the licences
didn't have matching IDs in the LGSL.

From oral history, we worked out that Dai made up prefixes for the
licences that weren't in the LGSL. He started at 9048 (for some reason),
and then added them sequentially all the way up to 9265.

So it makes sense for us to use 9266 as the prefix for our new licence.

Looking at the data in the database, the next digit after the dash is
always [1-8]. Most of the time it's 7. So that clearly has some meaning
too.

Again, from oral history we worked out that this is a region identifier:

1 - England
2 - Wales
3 - Scotland
4 - Northern Ireland
5 - England & Wales
6 - England & Wales & Scotland
7 - England & Wales & Scotland & Northern Ireland

(Aside: they've tried to cram four bits of information into a three bit
number here, so it's not possible to represent any combinations of two
nations [other than england and wales] or any combinations of three
countries [other than england, wales and scotland]. I guess this has
never come up in practice!?

There's one service in the database which has `8` as the middle number.
This looks like a mistake.

The final digit (after the second dash) is there to make sure the IDs
are unique. So if there are more than one licence for the same service,
targeting the same nations we can still have unique IDs. For our
purposes this can always be 1, because we're not trying to group
licences into services.

All of this was an interesting aside, but I'm pretty sure the fact that
the ID is meaningful is a huge red herring - there's nothing that parses
it, so it may as well just be a random UUID.

Software engineering lesson: don't imbue IDs with unnecessary meaning.
Use something you can guarantee will be unique (like a uuid, or an
incrementing integer), and store the meaningful bits somewhere else on
the record.

Now that we've worked all that out we can actually add the licence
entry. BEIS want it under Solicitor services / Offer professional legal
services, which we can do by adding a licence link.

I've made this task check if the licence is already there and error if
it is, so we don't need to remove it from the repo once it's run. It can
serve as an explanation of how to do this next time.

Because this is just a single licence, I've hardcoded everything in the
rake task. If we need to add batches of licences in the future we should
probably do something a bit better. There's no prior art in the git
history though.

Finally, once this has been added to licence-finder, we'll also need to
add a licence artefact in publisher
(https://publisher.publishing.service.gov.uk/artefacts/new). Publisher
will ask us for the ID, which will be 9266-5-1. This will end up in
search-api, which is how licence-finder turns our entry into a link to
the start page.