Fix search results when pagination happens
Previously `create_for_licences` made no attempt to paginate through
pages of search results. If you made a query which would return more
than the pagination limit, it would just return the first page and throw
away the rest.

The search API client has a search_enum method which does the pagination
for us. Calling `.to_a` forces it to go through all the pages and load
them into an array, which is consistent with what was there before.
There's no need to pluck out the `["results"]` anymore - that's an
artefact of a single page of results, so search_enum handles this for
us.

I believe this is only causing an issue on
www.gov.uk/licence-finder/browse-licences (which I added recently in
https://github.com/alphagov/licence-finder/pull/1004). That makes calls
to search-api 100 licences at a time, but each request only gets the
first page of results, so it makes it look like there are far fewer
results than there actually are.