Update jQuery from 1.12.4 to 3.6.3
Although there are many reasons to motivate an upgrade of such a legacy
dependency, the main motivation for this change was that jQuery 1.12
triggers CSP violations on initialisation in Firefox (due to [1] which
was removed by jQuery 2.2).

The reason this uses such an old version of jQuery is that we (GOV.UK)
used to share a version of jQuery across many apps and thus it was scary
to update as it was hard to test it's impacts. Since March 2022 [2]
Licence Finder has had it's own version of jQuery and thus it is easier
to judge the impacts.

To determine compatibly with jQuery 3 I ran this through jQuery migrate
[3] and fixed the one warning. I also went through and manually tested
the code setting breakpoints to check various parts were executed.

[1]: https://github.com/jquery/jquery/blob/e09907ce152fb6ef7537a3733b1d65ead8ee6303/src/event/support.js#L7-L20
[2]: https://github.com/alphagov/licence-finder/commit/c384fc507412d57cb8d312dcd3c285c98340fc29
[3]: https://github.com/jquery/jquery-migrate