Export level 2 sectors
Exports the level 2 sectors that are tagged to a licence.
Neither the level 1 and level 3 sectors nor the activities will be used
to tag licences in the new Find Licences finder, so they do not need to
be imported.

My understanding of how they models fit together is this:
Licences do not have a direct relationship to activities and sectors,
but instead have relationships with the LicenceLink model. It is assumed
that a Licence can have many LicenceLinks.

LicenceLinks belong to one licence, one activity and one sector. I think
that LicenceLinks only belong to level (layer) 3 sectors, but cannot be
sure, that is why a safe navigation operator is being used in the guard
clause, as the level (layer) 1 sectors do not have parents.

The query in the `level_two_sectors` method attempts to nicely replicate
the following database queries:

```
Sector.in(id: LicenceLink.where(licence_id: Licence.last.id).pluck(:sector_id)).last.parents.to_a
```

which returns something like:

```
[#<Sector _id: 4fec65542c9e04357e000143, public_id: 323, correlation_id: 1000325, name: "Legal services", layer: 2, parent_ids: [BSON::ObjectId('4fec65542c9e04357e000124')], activity_ids: []>]
```

but loops around all of the sectors.