Export licence data as API
As with the "browse sectors" end point, the API won't be indexed by
search, so therefore won't appear in Google searches or the GOV.UK site
search.

The data was initially exported in YAML format, but the data ended up
being 4,746 lines long on integration. This is a lot of data to copy and
format into a YAML file locally. Also, the timestamp needed to be
removed from the log of the YAML data, which is an added hassle.
Having a JSON endpoint removes this extra work.

Example output:

```json
[
  {
    "licence_identifier": "1-1-234",
    "locations": [
      "england",
      "wales",
      "scotland",
      "northern-ireland"
    ],
    "industry_sectors": [
      "motor-vehicle-and-fuel-retail-hire-and-repair"
    ]
  }
]
```

The api endpoint can be viewed locally at:
http://licence-finder.dev.gov.uk/licence-finder/licences-api